Update file names and move types to a separate folder
