+++ src/utils/types/RuleParamsBase.ts
